move volume class to own file
